refactor(isDateToday): use isSameDate - переписал isDateToday на переиспользование isSameDate - написал тестов
